requirements.txt
streamlit==1.28.0
pandas==2.1.3
python-docx==0.8.11
openpyxl==3.1.2
lxml==4.9.3
tzdata==2023.3